Add GameSense server ping with exponential backoff
The client now pings the GameSense server during the "probing" state to
determine if it should continue communication.

This also switches to using the FEvent interface instead of Promises and
Future objects for barriers since our use of promises left a lot of room
for potential crashes due to "broken" promises.